+++ webpack.config.js
@@ -1,5 +1,6 @@
 const path = require( 'path' );
 const UglifyJsPlugin = require( 'uglifyjs-webpack-plugin' );
+const WebpackAutoInject = require( 'webpack-auto-inject-version' );
 
 const env = process.env.WEBPACK_ENV;
 const plugins = [];
@@ -9,6 +10,22 @@ if ( env === 'min' ) {
   plugins.push( new UglifyJsPlugin() );
   ext = '.min.js';
 }
+plugins.push(
+  new WebpackAutoInject({
+    SHORT: 'textile-js | https://github.com/GehDoc/textile-js',
+    PACKAGE_JSON_PATH: './package.json',
+    components: {
+      InjectAsComment: true
+    },
+    componentsOptions: {
+      InjectAsComment: {
+        tag: 'Build version: {version} - {date}', // default
+        dateFormat: 'isoDateTime',
+        multiLineCommentType: false // default
+      }
+    }
+  })
+);
 
 module.exports = {
   mode: 'production',